voice/requirements.txt
SpeechRecognition==3.10.1
pyaudio==0.2.14
requests==2.31.0
websocket-client==1.6.4
